Adds traditional worker pool implementation
This breaks up the work into pairs of integers and distributes that
among the workers, with each worker pulling as many pairs off as it can
handle.

The earlier implementation specifically chunked up the work based on the
number of workers, so each worker in that implementation would only ever
process one job.

This new implementation is a more traditional worker pool design.

Notes:
- jobs are set as pairs of integers - practically, we would want to break
this up, probably based on a heuristic of the length of nums and count
of workers to reduce channel ops